+++ CHANGELOG.rst
@@ -5,6 +5,16 @@ codeaffen.phpipam Release Notes
 .. contents:: Topics
 
 
+v0.3.0
+======
+
+New Modules
+-----------
+
+- codeaffen.phpipam.domain - Manage L2 routing domains
+- codeaffen.phpipam.nameserver - Manage nameservers
+- codeaffen.phpipam.vlan - Manage vlans
+
 v0.2.0
 ======